Refactor main.rb file to align with SOLID
