+++ main.rb
@@ -1,14 +1,17 @@
 # rubocop:disable Metrics/CyclomaticComplexity
 
 require_relative './app'
+require_relative './menu'
 
 def main
   app = App.new
+  menu = Menu.new
+
   puts '
-    Welcome to School Library App!'
+  Welcome to School Library App!'
 
   loop do
-    display_options
+    menu.display_options
     choice = gets.chomp.to_i
 
     case choice
@@ -22,19 +25,6 @@ def main
                 exit
     end
   end
-end
-
-def display_options
-  puts '
-    Please choose an option by entering a number
-
-      1 - List all books
-      2 - List all people
-      3 - Create a person
-      4 - Create a book
-      5 - Create a rental
-      6 - List all rentals for a given person id
-      7 - Exit'
 end
 
 main
